Terminal output improvement
Checks were added to verify the terminal width is big enough to display
elements next to each other. This includes the order change output,
which is most likely to use a lot of horizontal space. Order changes
output lists will either be displayed as a single line if they consume
less than 50 % of the terminal width, or multi line otherwise.