Support order change detection in simple lists
Added code to detect order changes in simple lists based on the hash
of the respective entries. The output currently displays the order of
the entries that are common to both lists. This might be subject to
changes in the future if this turn out to be too much data to be
displayed.