Reworked error handling
The `dyff` package should never panic or exit the program, but only
return an error that can be used by an internal end-user centric
package. Therefore, the `ExitWithError` function was moved out
of the `dyff` package and all functions that could error will return
an error as a secondary return value.