Moved the banner code into the output source file
The main tool subcommands should be as simplistic as possible. That is
why the banner output and logic around that was moved into the `dyff`
package. This change includes an improvement on how to decide whether
the document index has to be displayed in the path. Also, the location
information now comes with a hint on how many documents are in the input
file.