Implement change root support
Add flags and implementation code to support changing the root level
of a given input file, or both. Depending on the actual object that
is referenced by the path that serves as the new root level, there
are use cases where a list should be translated into documents, or
where the list is used as the root level itself.