Support non-standard identifier in named lists
Add function to guess the identifier if the default identifier name,
key, or id do not work. It will guess based on the fact if there is
a string key with string value, where each value is unique in the
list as long as the list has more than `n` values with `n` currently
being 3. This threshold might be subject to change in the future or
become configurable.