Add `in-place` flag
Add support for flag `--in-place` for JSON and YAML output. This
outputs the result to original file replacing it in place. The flag
makes sense in combination with `--restructure` in order to re-order
flags of an existing file.

The commit includes a rework of the output function code in order
to avoid existing code duplication.

This fixes issue #10.