Fix identifier detection in named-entry lists
Rework `getIdentifierFromNamedLists` to check not just the number of keys in
the list, but more precisely the number of unique entries. This is to detect
lists that have a `name` key in each entry that does not have unique values.

Improve flaky test where the path style rendering option was not reset
correctly for each new test section.

This commit should fix #38.