Introduce command settings and RunE
At the moment, a lot of global variables are used for the flag
processing in dyff and they are even shared between commands. This works
like it is, but it is not best practice and blocks improved testing of
the commands package.

Introduce `RunE` instead of `Run` for all Cobra commands that are
targeted for later testing to have a consistent interface to work with.

Add command settings structs to encapsulate the respective command flag
settings in one data structure. Include a `ResetSettings` function to
have one place to reset all settings to factory default for tests.