Rewrite cmd package unit test to use program entry
The test cases in the command package did not use the main program entry
function, but used the internals directly. Therefore, this did not test
the commands package, but internals in the core package.

Introduce wrapper functions to run `dyff` from inside the test framework
in a way that is comparable to an individual running dyff binary.