Fix indention issue in type change output
As described in issue #89, the `dyff` output uses a wrong indention if
the type change shows a multi-line difference, for example in case the
original type is a mapping context.

Output looks like this:
```
foo
  ± type change from map to list
    - a: 1
  b: 2
    + - 1
  - 2

bar
  ± type change from map to <nil>
    - c: 3
  d: 4
    + <nil>
```

Instead, it should look like this:
```
foo
  ± type change from map to list
    - a: 1
      b: 2
    + - 1
      - 2

bar
  ± type change from map to <nil>
    - c: 3
      d: 4
    + <nil>
```

Use `createStringWithPrefix()` in the type change output code, which
supports multi-line output already.