Add option to omit dyff banner header
Reported with #105, it would be helpful to have an option to omit the
default `dyff` banner header with the report summary and the file names.

Add flag `--omit-header` to disable the banner header.

Refactor the command settings structs to avoid duplicate code in the
reset logic for the test cases.

Add test case to verify the command flag omits the dyff banner header.