Add ignore order changes flag
Based on feedback in #108, there are use cases where the order changes
in a list are not important to be reported. It should be possible to
ignore changes to the order of a list.

Add flag to ignore order changes for the `between` command.

Add new test case to verify new command flag.

Add global option in `dyff` package to disable order change detection.

Introduce short circuit to bypass the order change detection.

Refactor return statement for named entry lists to unify the style.