Refactor global varibales for diff options
The usage of global variables makes it harder to test the package and is
generally not the best pattern for non constant settings.

Introduce compare options that can optionally be used during the compare
to specify the way the comparison is done, for example whether to ignore
order changes in lists.

Move style settings for Go Patch style and minor change threshold into
the report source file.

Move code that is only used for the human report into the respective
source code file.

Introduce test case to show that the identification of non-standard
named entry list identifiers can fail if configured differently.

Refactor code to match a more common stlye when it comes to multiline
instructions with lots of parameters.