Fix panic when using timestamps in output report
As reported by #111, in case a change of a timestamp value needed to be
reported, dyff failed with a panic. The code was intended to find types
that were not supported.

Change human readable type code to be more generic to support all
possible types by using the YAML tag name as the default.