Introduce last-applied command
When working with `kubectl` and Kubernetes YAML files, an update to a
configuration can result in an additional metadata header, which stores
the previous configuration. Since this is a string containing JSON, it
is not easy to spot the difference immediatly.

Add new command calle `last-applied`, which takes in one location
argument, for example Stdin (`-`) or a file location. In case the well
known metadata header is present, a difference report of the previous
configuration and the current configuration is created.

Refactor report options into a seperate variable to that it can be
shared between different commands.

Add function to apply report option related flags to a command.

Rename `ShowBanner` to `OmitHeader` to harmonise the flag settings.

Bump `ytbx` to latest available version to get the `Delete` function.