Refactor banner creation code
The code for creating the banner header was written overly complex and required
additional helper functions to place everything at the correct spot.

Switch to an implementation that uses only one string for the banner header.

Use new `bunt` package feature to colorize the tool name in the banner.

Bump dependencies to latest available versions.