Sort simple list entry for hash calculation
Fixes #125

The `--ignore-order-changes` flag enabled specific code sections to search for
order changes in simple and named entry lists. However, there is also the
possibility that `dyff` should ignore the order in lists in case it compares
two entries of a simple lists that are lists themselves. For this use case the
normal logic does not apply and does not produce the expected result. Currently
both entries result in different hashes based on their list entry order.

Add sorting function to sort list entries before creating a hash.