Reports can be filtered on set of paths
Fixes #157

User wants to focus on specific differences and therefore wants to filter on
paths in the report. All the others should be omitted.

Refactor `Filter` function to make use of Map for more efficient
processing.

Add flag for provinding filtering arguments.

Add `reports.go` for methods of `Report` structure

Signed-off-by: Baran Dalgic <baran.dalgic@posteo.net>