Enable path to indicate file level
The `path` is used to show the path in which a difference is located. It can
show the path in the document and the document identifier. However, it cannot
show a different at the file level, where the documents are stored.

Make path optional (as pointer) to enable file level differences.