Enable the addition of custom identifier keys
To compare maps on complex YAMLs, the library requires that
items are compared by finding a key to match elements on.  There
is a heuristic list of identifying keys that works in most cases.

This change enables additional keys to be explicitly added for
YAMLs that follow a known schema.