feat: allow kubernetes generateName
Kubernetes provides the ability to create objects with generateName, which acts as a prefix and the suffix is added by the api-server and returned to the client.  I discovered while attempting to diff my manifest using dyff that this seems unsupported at the moment.

I'm proposing adding a new flag here to maintain backwards compatibility that will enable matching on generateName.